move wlr_box from /types to /util
