+++ include/wlr/render/interface.h
@@ -13,9 +13,11 @@
 #include <wayland-server-core.h>
 #include <wlr/render/wlr_renderer.h>
 #include <wlr/render/wlr_texture.h>
-#include <wlr/types/wlr_box.h>
 #include <wlr/types/wlr_output.h>
 #include <wlr/render/dmabuf.h>
+
+struct wlr_box;
+struct wlr_fbox;
 
 struct wlr_renderer_impl {
 	bool (*bind_buffer)(struct wlr_renderer *renderer,
