+++ include/wlr/types/wlr_cursor.h
@@ -10,7 +10,6 @@
 #define WLR_TYPES_WLR_CURSOR_H
 
 #include <wayland-server-core.h>
-#include <wlr/types/wlr_box.h>
 #include <wlr/types/wlr_input_device.h>
 #include <wlr/types/wlr_output_layout.h>
 #include <wlr/types/wlr_output.h>
@@ -26,6 +25,7 @@
  * the screen, and so on.
  */
 
+struct wlr_box;
 struct wlr_cursor_state;
 
 struct wlr_cursor {
