+++ include/wlr/types/wlr_pointer_constraints_v1.h
@@ -12,7 +12,6 @@
 #include <stdint.h>
 #include <wayland-server-core.h>
 #include <pixman.h>
-#include <wlr/types/wlr_box.h>
 #include <wlr/types/wlr_seat.h>
 #include "pointer-constraints-unstable-v1-protocol.h"
