+++ include/wlr/types/wlr_xdg_shell.h
@@ -8,9 +8,10 @@
 
 #ifndef WLR_TYPES_WLR_XDG_SHELL_H
 #define WLR_TYPES_WLR_XDG_SHELL_H
-#include <wlr/types/wlr_box.h>
-#include <wlr/types/wlr_seat.h>
+
 #include <wayland-server-core.h>
+#include <wlr/types/wlr_seat.h>
+#include <wlr/util/box.h>
 #include "xdg-shell-protocol.h"
 
 struct wlr_xdg_shell {
