+++ render/wlr_renderer.c
@@ -5,6 +5,7 @@
 #include <wlr/render/pixman.h>
 #include <wlr/render/wlr_renderer.h>
 #include <wlr/types/wlr_matrix.h>
+#include <wlr/util/box.h>
 #include <wlr/util/log.h>
 
 #include <wlr/config.h>
