+++ types/wlr_pointer_constraints_v1.c
@@ -4,9 +4,9 @@
 #include <stdbool.h>
 #include <stdlib.h>
 #include <wayland-server-core.h>
-#include <wlr/types/wlr_box.h>
 #include <wlr/types/wlr_pointer_constraints_v1.h>
 #include <wlr/types/wlr_region.h>
+#include <wlr/util/box.h>
 #include <wlr/util/log.h>
 #include "util/signal.h"
