+++ util/meson.build
@@ -1,5 +1,6 @@
 wlr_files += files(
 	'array.c',
+	'box.c',
 	'global.c',
 	'log.c',
 	'region.c',
